various: add autoreview topic assignment to autosubmit label users
try_set_autosubmit_labels is often used as a convenience function for
"make it as easy as possible for the reviewers to land this." These
kinds of CLs are great for automatic pings on reviews, too.

Since these are migrating to this new flow, also remove the existing
reviewers that were selected for these CLs.

BUG=b:409821544
TEST=repo upload

Change-Id: I537e0baf4c8d57df2662fd7092ff9d28a965da17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6448544
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Saketh Pothireddy <spothire@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>