remove final check-presubmit mypy blocklist entry
`crate_ebuild_help.py` creates ebuilds in a style that we haven't
encouraged for quite a while: (using `CRATES` instead of
`third-party-crates-src`). Further, it has a `TypeError` that should be
breaking most attempts to use it:

```
with TemporaryDirectory() as foo:
  # `.name` only exists on the `TemporaryDirectory`; if you use it as
  # a context manager, `foo` is a `str`, so this is unconditionally
  # broken.
  bar(foo.name)
```

Remove it and see if anyone yells. It can be easily restored if there're
still users.

BUG=b:413369761
TEST=None

Change-Id: I83f55375ebb8e5542574cdaca903d1e789dc242e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6498607
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>