add run_weekly_job
This is a simple script that manages the running of a subcommand.
Essentially, the goal boils down to "if the script-to-run is successful,
run it weekly. Otherwise, run it more often than that. Also ignore the
first few failures, because it could be flakes."

There's already logic that's basically this for Cr-Os jobs, and I'll
soon be introducing more similar jobs; might as well centralize the
logic.

Given the simplicity of this check and this repo's `mypy` requirements,
I didn't feel like writing tests would be useful. Happy to refactor and
add a few if someone wants.

BUG=b:420954566, b:416321626
TEST=Ran in various configurations; all WAI.

Change-Id: Id611e206af0b24bb351e35d2ecc649e70f3b5770
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6612019
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>