afdo_tools: remove run_afdo_tryjob
This fails to start on python>=3.13 due to the removal of the `pipes`
import.

This could easily be fixed, but this script seems entirely unused in the
last 5 years, so there's no clear point in continuing to maintain it.

BUG=b:447150674
TEST=None

Change-Id: Ie7f013bf382228eb5c9e17efd27490354f1622ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/6980385
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Saketh Pothireddy <spothire@google.com>
Auto-Submit: George Burgess <gbiv@chromium.org>
Commit-Queue: Saketh Pothireddy <spothire@google.com>