llvm_tools: save Gemini cache after check_revert failures
Gemini failed to perform inference correctly on a commit up to
check_reverts.py's retry limit, leading to ~6K inference results being
dropped on the floor.

Each inference by this tool is quite cheap, but there _are_ daily
request limits to be mindful of, and ~6K inference results took an hour
to collect with default job counts.

Add a bit of logic so we're _less likely_ to drop successful inference
results in the future.

Gemini was used to generate the tests, and I tweaked them somewhat.

BUG=b:436267619
TEST=Ran script over 28K commits, got rate-limited after 3K, second
TEST=run attempt tried to fetch 25K commits

Change-Id: I6ae46919904c595c688908aedbe6383f01fcc842
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7000838
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>