android_tools: add warning suppression patch cleanup script
This allows us to use Gemini for cleaning up warning suppression CLs.
b/445908399#comment35 indicates that its capabilities need improvement
(gemini's edits are 100% correct, but it only makes edits in 36% of the
cases that it ideally should).

That said, this is a foundation that can be built on in a few ways, and
there's obvious area for cleaning up the diff in this script without
Gemini, so I'm opting to send for review as-is.

BUG=b:445908399, b:455912162
TEST=Ran on art CL; all edits were correct, it'd just be nice if it did
TEST=more edits...

Change-Id: I08e466799b6fb139b03771dcf13fd04d46e8c305
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7096293
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>