llvm_tools: fix HEAD staleness bug
This bookkeeping has to happen on every call to this function, but the
`if not new_reverts ... continue` will cause it to be skipped.

Move this into its own loop, so it's harder to accidentally skip this
logic. Move the loop into a function so this can be easily tested.

Test added by Gemini, reviewed by me.

BUG=b:458341903
TEST=Unittests

Change-Id: I35002bf5c7756b802d9c32e26ce4be11a6c947d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7138096
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>