llvm_tools: ignore reverts of old commits
This is a partially-complete solution to this problem, and it should
filter out all of the most obvious pieces of noise.

It's not necessarily that reverts of very old commits are _worthless_,
just they've very likely already survived an LLVM release, so they
don't seem likely to be high signal. Happy to be proven wrong, though.

Tests written by Gemini.

BUG=b:458341903
TEST=Unittests

Change-Id: If482fdc85b18a4b325f57fbcb447a0901561e427
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7138097
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>