llvm_tools: Use topic for LLVM reverts, query them with Gerrit
This updates the logic for nightly_revert_checker and
generate_llvm_revert_report, to fix b:461904503 where many reverts are
not included in the report.

This makes the following changes:

1. nightly_revert_checker now adds topic "revert-checker" on the CLs
   it uploads.

2. generate_llvm_revert_report now queries Gerrit for changes that match
   the topic and the given branch.

The output of generate_llvm_revert_report has also changed; it now has
columns Status, URI, Subject, and Notes.

This output can be copy-pasted into a spreadsheet, and gives an overview
of all reverts that target a given LLVM branch, with a status that
indicates if they are MERGED, ABANDONED, or NEW (the last indicating
the CL needs the mage's attention).

BUG=b:461904503
TEST=unittest, also ran it for chromeos/llvm-r574158-1

Change-Id: I14c8a30413a70dce1253c02b0c5f837b412bd474
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7170958
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>