fix up existing `__name__ == "__main__"` issues
The tl;dr of the issue here is "these are just noise in the
overwhelmingly common case for toolchain-utils, because we invoke
everything through wrappers, so `__name__ == "__main__"` is almost
never True."

For 2 of these cases, `__name__ == "__main__"` is pointless noise.
For the `gemini_api` one, there was a consistency check that initially
worked fine, but once that was migrated to our standard wrapper flow,
that check got silently dropped.

These were identified by grepping for `__name__ == "__main__"` and
running presubmits.

Gemini handled the fixes, and I reviewed + readded the `_verify` call
that was dropped.

BUG=b:468297046
TEST=repo upload

Change-Id: I207481ba4d56c6039b7c689e8484d79caac2eea6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7256470
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>