pyproject: set strict_equality for mypy
Few cleanups needed, but they seem to be bugs, since

```
>>> pathlib.Path('/') == '/'
False
```

Cleanups by Gemini, though the refactor in get_patch was by me. The bug
there was `Path.root` is a `str`, and repeated `resolve()` actions were
a bit odd. Since we have an existing hammer for this (that should
probably be in cros_paths...), just reuse that.

BUG=b:468297046
TEST=repo upload

Change-Id: I955342df85f469f66e92923984b6dc887c4d9f6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/7262402
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: George Burgess <gbiv@chromium.org>